+++ millennia/generate_tables.py
@@ -1,4 +1,3 @@
-import re
 from collections import OrderedDict
 from operator import attrgetter
